+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## master
 
+ * Add the compact mode [#125](https://github.com/railsware/js-routes/pull/125)
+ * Routes path specs [#135](https://github.com/railsware/js-routes/pull/135)
+
 ## v0.9.9
 
 * Bugfix Rails Engine subapplication route generation when they are nested [#120](https://github.com/railsware/js-routes/pull/120)